+++ c/containers/slls.h
@@ -17,7 +17,7 @@ typedef struct _sllse_t {
 typedef struct _slls_t {
 	sllse_t *phead;
 	sllse_t *ptail;
-	int      length;
+	unsigned long long length;
 } slls_t;
 
 slls_t* slls_alloc();
